Sun, Moon, Stars, optimized chunkRenderPipeline
Added celestial objects that move as time passes. Moving clouds. Implemented .json profiling/scoping. And the best thing is that before every rendered chunk had 2vbos and 2vaos, now all chunks share a single vbo and vao -> like that's awesome, innit? And last but not least Fixed bug of using vSync - it has been turned off, so now the pipeline doesn't need wait each frame and pause the whole main thread. Also NVIDIA GPU will be prefered in contrast to  integralGPU

+++ Sandbox/src/world/entity/entities.h
@@ -190,7 +190,7 @@ public:
 	virtual ~EntityRoundBullet() = default;
 
 	EntityType getEntityType() const override;
-	std::pair<int, int> getLightPosition() const override { return std::make_pair(m_pos.x, m_pos.y); }
+	std::pair<int, int> getLightPosition() const override { return std::make_pair((int)m_pos.x, (int)m_pos.y); }
 	uint8_t getIntensity() const override { return 10; }
 	bool onEntityHit(World& w, WorldEntity* entity) override;
 	bool onBlockHit(World& w, int blockX, int blockY) override;
